feat: create separate AuthorityEvent component
