+++ resources/views/components/authority-event.blade.php
@@ -1,0 +1,3 @@
+@props(['event', 'count' => 0, 'i' => 0])
+<strong><a
+        href="{!! route('frontend.author.index', ['place' => $event->place]) !!}">{!! $event->place !!}</a></strong>{{ $formatEvent($event) }}{{ $i + 1 < $count ? ', ' : '' }}
